requirements.txt
numpy>=1.24.3,<2.0.0
fastapi
uvicorn[standard]
streamlit
python-dotenv
requests
pymupdf
pdf2image
pillow
pytesseract
transformers
accelerate
sentence-transformers[cross-encoder]
faiss-cpu
scikit-learn
docx2txt
TTS==0.21.3
gTTS
faster-whisper
python-multipart
